requirements.txt
pillow
PySimpleGUI